+++ CHANGELOG.md
@@ -31,6 +31,9 @@
 * Action when Set record for deletion option is invoked. Refs UIIN-2595.
 * Detail view not opened for non-local items when one shared record found using "Barcode" search on "Item" tab. Fixes UIIN-2698.
 * Jest/RTL: Cover CreateHoldings component with unit tests. Refs UIIN-2663.
+* Display Holdings/Item Electronic Access URIs consistent with Instances. Fixes UIIN-2778.
+* Staff suppress facet - use `No` as default value. Hide the facet when users don't have permissions to use it. Refs UIIN-2596.
+* Add a new search option for instances called `LCCN normalization`. Refs UIIN-2245.
 * Add and adjust collapse/expand buttons for consortial instances. Refs UIIN-2711.
 
 ## [10.0.10](https://github.com/folio-org/ui-inventory/tree/v10.0.10) (2024-01-17)